Relate Recipe to User using the "Recipe Owner" Join table
Exposing the join table as a resource in its own right is apparently the correct thing to do here? It feels weird that all my database entities are being mapped directly to the API but I found a stackoverflow that supported this using rationale that the relationship itself may have salient column information in the future (which it totally will, "created" boolean flag).

I think this API is progressing a little bit, now the next piece is ingredients.